improvements, refactoring
